This is everything we did together in the MC plus a few things I changed after you left. The "blanks" list is now only found inside the "Puzzle" class. I made a new method to show the current word and changed "newWord" to "currentWord". I think the CheckWinner function is going to either be rethought or moved to be inside the Puzzle class.
This is everything we did together in the MC plus a few things I changed after you left. The "blanks" list is now only found inside the "Puzzle" class. I made a new method to show the current word and changed "newWord" to "currentWord". I think the CheckWinner function is going to either be rethought or moved to be inside the Puzzle class.